chore: refactor and separate UI components for better modularity
