+++ src/App.tsx
@@ -1,11 +1,20 @@
-import React from 'react';
 import { ThemeProvider } from './context/ThemeContext';
 import DrawingCanvas from './components/DrawingCanvas';
+import { PanelProvider } from './context/PanelContext';
+import { PanelIOProvider } from './context/PanelIOContext';
+import { CanvasSettingsProvider } from './context/CanvasSettingsContext';
+
 
 function App() {
   return (
     <ThemeProvider>
-      <DrawingCanvas />
+      <CanvasSettingsProvider>
+        <PanelProvider>
+          <PanelIOProvider>
+            <DrawingCanvas />
+          </PanelIOProvider>
+        </PanelProvider>
+      </CanvasSettingsProvider>
     </ThemeProvider>
   );
 }
